split trajproc.py into multiple seperate analysis files and reorganise folders
